requirements.txt
kafka-python==2.0.2
pytest
pytest-mock
coverage